+++ app/models/index.js
@@ -5,7 +5,7 @@ var path      = require("path");
 var Sequelize = require("sequelize");
 var basename  = path.basename(module.filename);
 var env       = process.env.NODE_ENV || "development";
-var config    = require(__dirname + '/../config/config.json')[env];
+var config    = require(__dirname + '/../../config/config.json')[env];
 var sequelize = new Sequelize(config.database, config.username, config.password, config);
 var db        = {};
 
@@ -23,6 +23,11 @@ Object.keys(db).forEach(function(modelName) {
   if ("associate" in db[modelName]) {
     db[modelName].associate(db);
   }
+});
+
+var migrator = sequelize.getMigrator({
+  path:        process.cwd() + '/../../migrations',
+  filesFilter: /\.js$/
 });
 
 db.sequelize = sequelize;
