+++ app/models/user.js
@@ -1,7 +1,7 @@
 var bcrypt = require('bcrypt');
 
 module.exports = function(sequelize, DataTypes) {
-    var User = sequelize.define("User", {
+    var User = sequelize.define('User', {
         id:{ type: DataTypes.INTEGER, autoIncrement:true, allowNull:false, primaryKey:true},
         username: DataTypes.STRING,
         age: DataTypes.INTEGER,
