+++ app/views/added.jade
