+++ app/views/deleted.jade
