+++ app/views/footer.jade
