+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "courseby-node-express",
+  "version": "1.0.0",
+  "description": "An Express.js app for course.by",
+  "main": "index.js",
+  "dependencies": {
+    "bcrypt": "^0.8.0",
+    "body-parser": "^1.10.0",
+    "express": "^4.10.6",
+    "jade": "^1.8.2",
+    "sequelize": "^2.0.0-rc1",
+    "sequelize-cli": "^1.0.1",
+    "sqlite3": "^3.0.4"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/okendoken/courseby-node-express.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/okendoken/courseby-node-express/issues"
+  },
+  "homepage": "https://github.com/okendoken/courseby-node-express"
+}
